+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 3.13.2
+## *Fix*
+- synchronous call to delete_user_permission_project
+
+# 3.13.1
+## *Hotfix*
+- remove send_message from tasks
+
+# 3.13.0
+## *Add*
+  - project description
+
+## *Update*
+  - project description sended on project create and update
+
 # 3.12.0
 ## *Remove*
   - Remove project authorization list from api return